some refactoring: outsourced home component to own file
