+++ src/components/App/App.js
@@ -1,5 +1,3 @@
-import { useState } from 'react'
-import styled from 'styled-components'
 import { nanoid } from 'nanoid'
 import React from 'react'
 import {
@@ -10,20 +8,8 @@ import {
   Routes,
 } from 'react-router-dom'
 
-import { Todo } from '../TaskObjects/Todo'
-import { Add } from '../TaskObjects/Add'
 import { Header } from '../Header'
-import useTodo from '../../common/useTodo'
-import shallow from 'zustand/shallow'
-
-const defaultTask = [
-  {
-    id: nanoid(),
-    task: 'write new task',
-    completed: false,
-    archived: false,
-  },
-]
+import { Home } from '../Home'
 
 function App() {
   return (
@@ -40,26 +26,6 @@ function App() {
   )
 }
 
-function Home() {
-  const todoIds = useTodo(
-    (state) =>
-      state.todos.filter((todo) => !todo.archived).map((todo) => todo.id),
-    shallow
-  )
-  console.table(todoIds)
-  return (
-    <>
-      <Body>
-        <Add />
-
-        {todoIds.map((id) => {
-          return <Todo key={id} id={id} />
-        })}
-      </Body>
-    </>
-  )
-}
-
 function Archive() {
   return <h2>Archive</h2>
 }
@@ -67,13 +33,5 @@ function Archive() {
 function Random() {
   return <h2>Random</h2>
 }
-
-// TODO auslagern
-
-const Body = styled.body`
-  height: 90vh;
-  background: #282c34;
-  padding: 10vw;
-`
 
 export default App
